little refactoring, still no luck with the get station call.

+++ Pandora.Api/Dtos/BaseRequestDto.cs
@@ -10,8 +10,8 @@ namespace Pandora.Api.Dtos
     {
         public string PartnerAuthToken { get; set; }
         public string PartnerId { get; set; }
-        public double PartnerRequestSyncTime { get; set; }
-        public double PartnerResponseSyncTime { get; set; }
+        public int PartnerRequestSyncTime { get; set; }
+        public int PartnerResponseSyncTime { get; set; }
         public string UserId { get; set; }
         public string UserAuthToken { get; set; }
         public string EncryptionKey { get; set; }
